move HA type declaration to common + update imports
